requirements.txt
gradio>=4.31.0
sqlalchemy>=2.0.25
bcrypt>=4.1.2
PyJWT>=2.8.0
python-dotenv>=1.0.1
httpx>=0.27.0
openai>=1.35.0
google-api-python-client>=2.130.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
